Implement mock backend
The mock backend allows users to try the webapp
without actually connecting to a real gRPC backend.
The idea is to host the webapp as a github page, with
mocked backend enabled by default.

Telemetry is mocked as best effort, with simulated data
for all telemetry frames (battery, attitude, gps, baro, vario, etc).
The simulated data is not self-consistent. It's just meant to look
semi-realistic.

The list of gamepads, and list of transmitters is mocked properly.

The link stream, and link state is mocked properly.

The channels, and gamepad input states streams are not fully mocked yet.
The values stay fixed. The idea is to eventually mock gamepad input
values, and actually evaluate the config within JavaScript to determine
channel values.

The eval data stream, is not yet implemented.