Fix USB enumeration for Raspberry Pi 4
When testing on Raspberry Pi 4, only
one USB device was showing whe using two sticks of the same
brand, even though they had different names.

This was becaue the GUID being reported by underlying OS was the same
for both sticks.

This change modifies the logic such that the name is also included
when computing the stick hash/ID. Even then, there is still a chance
that you can connect two identical sticks. In that case, there
is additional logic where the enumeration index is appended to the hash/ID
to distinguish between the two.